feat: separation of concern
